refactor: add useDatepickerState hook and move types to separate file
