+++ src/components/Datepicker/Datepicker.tsx
@@ -1,14 +1,11 @@
-import React, { FC, useRef, useReducer } from 'react';
+import React, { FC, useRef } from 'react';
 import { DatepickerWrapper } from './styled';
 import Global from '../../styles/Global';
 import Input from '../Input/Input';
 import Calendar from '../Calendar/Calendar';
-import { Mode } from '../../utils/ClassCalendar';
 import { useOutsideClick } from '../../hooks/useOutsideClick';
-
-export type View = 'weekly' | 'monthly';
-
-export type WeekendStatus = 'show' | 'hide';
+import { useDatepickerState } from '../../hooks/useDatepickerState';
+import type { Mode, View, WeekendStatus } from '../../types/types';
 
 export interface DatepickerProps {
   start: Mode;
@@ -18,51 +15,8 @@ export interface DatepickerProps {
   minDate: Date;
 }
 
-export const init: DatepickerProps = {
-  start: 'mo',
-  view: 'monthly',
-  weekend: 'show',
-  minDate: new Date(2021, 3, 2),
-  maxDate: new Date(2025, 4, 25),
-};
-
-export interface Init {
-  from: Date | null;
-  to: Date | null;
-  currentCalendar: 'From' | 'To';
-  showCalendar: boolean;
-}
-
-const initialValue: Init = {
-  from: new Date(2022, 5, 17),
-  to: new Date(2023, 7, 17),
-  currentCalendar: 'From',
-  showCalendar: false,
-};
-
-export type Action =
-  | { type: 'SET_DATE_FROM'; payload: { date: Date | null } }
-  | { type: 'SET_DATE_TO'; payload: { date: Date | null } }
-  | { type: 'SET_CURRENT_CALENDAR'; payload: { calendar: 'From' | 'To' } }
-  | { type: 'TOGGLE_CALENDAR'; payload: { showCalendar: boolean } };
-
-const reducer = (state: Init, action: Action) => {
-  switch (action.type) {
-    case 'SET_DATE_FROM':
-      return { ...state, from: action.payload.date };
-    case 'SET_DATE_TO':
-      return { ...state, to: action.payload.date };
-    case 'SET_CURRENT_CALENDAR':
-      return { ...state, currentCalendar: action.payload.calendar };
-    case 'TOGGLE_CALENDAR':
-      return { ...state, showCalendar: action.payload.showCalendar };
-    default:
-      return state;
-  }
-};
-
 const Datepicker: FC<DatepickerProps> = ({ ...initial }) => {
-  const [state, dispatch] = useReducer(reducer, initialValue);
+  const { state, dispatch } = useDatepickerState();
   const calendarRef = useRef<HTMLDivElement>(null);
 
   const handleCloseCalendar = (value: boolean) => {
